added invoices endpoint to customercontroller

+++ src/main/java/com/eindopdracht/eindopdracht/repository/InvoiceRepository.java
@@ -4,5 +4,5 @@ import com.eindopdracht.eindopdracht.model.Invoice;
 import org.springframework.data.repository.CrudRepository;
 
 public interface InvoiceRepository extends CrudRepository<Invoice, Long> {
-    Iterable<Invoice> findByCustomerId(Integer customerId);
+    Iterable<Invoice> findByCustomerId(Long customerId);
 }
